pyproject.toml
[tool.poetry]
name = "nonebot-adapter-minecraft"
version = "1.0.8.post2"
description = "NoneBot2与MineCraft Server互通的适配器。"
authors = ["17TheWord <17theword@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "nonebot" }]

[tool.poetry.dependencies]
python = "^3.9"
nb-cli = "^1.4.0"
aio-mc-rcon = "^3.2.2"
nonebot2 = { version = "^2.2.0", extras = ["fastapi"] }


[tool.poetry.group.test.dependencies]
anyio = "~=3.6"
nonebug = "^0.3.5"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.data.packages]
my_package = { path = "nonebot/adapters/minecraft" }